refactor: Improve layout and UI responsiveness
Adjusted the `Layout` component to use `h-[100dvh]` and `overflow-hidden` for proper full viewport height and to prevent scrolling.
Modified the sidebar to be a fixed-width, full-height element on larger screens.

In `Sales.tsx`, simplified the logic for displaying the `priceAtSale` input value, defaulting to an empty string when the price is 0.
In `Clients.tsx` and `Suppliers.tsx`, removed the `useRef` for menu management and replaced it with a state variable `menuPos` to track menu position. Also, added scroll event listeners to close the dropdown menus, enhancing usability.
In `Expenses.tsx`, refined the `handleDeleteCategory` function to prevent deletion of default categories and provide clearer user confirmation.